clipboard-apis/async-html-script-removal.https.html

<!doctype html>
<meta charset="utf-8">
<title>
  Async Clipboard write ([text/html ClipboardItem]) -> readHtml tests
</title>
<link rel="help" href="https://w3c.github.io/clipboard-apis/#async-clipboard-api">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script>
'use strict';
// This function removes extra spaces between tags in html. For example, the following
// html: "<p> Hello </p>   <body> World </body>" would turn into this
// html: "<p> Hello </p> <body> World </body>"
function reformatHtml(html) {
  let parser = new DOMParser();
  let htmlString = parser.parseFromString(html, 'text/html').documentElement.innerHTML;
  let state = 0;
  let reformattedString = '';
  let textBetween = '';
  let justSpaces = true;
  for (let i = 0; i < htmlString.length; i++) {
    let c = htmlString.charAt(i);
    if (state == 0) {
      if (c == '>') { state = 1; }
      reformattedString += c;
    } else if (state == 1) {
      textBetween += c;
      if (c == '<') {
        if (justSpaces) {
          reformattedString += ' <';
        } else {
          reformattedString += textBetween;
        }

        textBetween = '';
        justSpaces = true;
        state = 0;
      } else if (c != ' ') {
        justSpaces = false;
      }
    }
  }
  return reformattedString;
}

const html_with_script = '<title>Title of the document</title> <script>const a = 5;</scr' + 'ipt> <p>Hello World</p>';
const html_without_script = '<title>Title of the document</title>   <p>Hello World</p>';
promise_test(async t => {
  await test_driver.set_permission({name: 'clipboard-read'}, 'granted');
  await test_driver.set_permission({name: 'clipboard-write'}, 'granted');
  const blobInput = new Blob([html_with_script], {type: 'text/html'});
  const clipboardItem = new ClipboardItem({'text/html': blobInput});
  await navigator.clipboard.write([clipboardItem]);
  const clipboardItems = await navigator.clipboard.read({type: 'text/html'});

  const html = clipboardItems[0];
  assert_equals(html.types.length, 1);
  assert_equals(html.types[0], 'text/html');

  const blobOutput = await html.getType('text/html');
  assert_equals(blobOutput.type, 'text/html');

  const blobText = await (new Response(blobOutput)).text();

  let outputHtml = reformatHtml(blobText);
  let inputHtml = reformatHtml(html_without_script);
  assert_equals(outputHtml, inputHtml);
}, 'Verify write and read clipboard given text: ' + html_with_script);
</script>
